Remove the precision bug from the C variant
This was introduced in d88c0c8f8d1effd5bd395cb71657629fb242f661 to mirror
Fortran reference results. The bug was fixed in the Fortran variants in
bac8d631f017a391ee8441309247eadd51dca637 and reference results updated
but the fix was not applied to the C variant at that time.