Loki-SCC/H: Add yoecldp.F90 header to ensure correct array demotion
The crucucial piece of information in this header file is the fact
that is known (NCVL=5), which allows us to treat arrays with this
dimension as otherwise scalar. The Loki-SCC/H transformations will
pick up on that fact, but need to know the type information of that
integer parameter, which they cannot otherwise infre.