Add semantic for negative thread count
Zero or negative values given as first argument of the various prototypes
led to errors. This commit adds semantic to 0 or negative values by enabling
the program to read the OMP_NUM_THREAD value from the environment, or using
the number of procs availables if the variable is not defined.