Loki: Only use stmt functions via imported headers in Loki variants
This removes a lot of the handholding needed for the C-specific
transpile variant, where the utility functions where previously
inlined from re-written elemental functions. Instead we now configure
the Loki `InlineTransformation` to inline those, and force the CLOUDSC
kernel to be preprocessed, so that we may find the comdeck functions.

Note, I've left the source for the re-written elemental functions in
the common directory for reference. If that becomes annoying, these
can now be removed.